+++ src/lib/export_abi.ts
@@ -1,22 +1,13 @@
-import pkg from '../package.json';
-import { AbiExporterConfig, AbiExporterConfigEntry } from './types.js';
+import pkg from '../../package.json';
+import { AbiExporterConfig, AbiExporterConfigEntry } from '../types.js';
+import { clearAbi } from './clear_abi.js';
 import { FormatTypes, Interface } from '@ethersproject/abi';
-import deleteEmpty from 'delete-empty';
 import fs from 'fs';
 import { HardhatPluginError } from 'hardhat/plugins';
 import { HookContext } from 'hardhat/types/hooks';
 import path from 'path';
 
 const TS_TAG = `// this file was automatically generated by ${pkg.name} - do not modify`;
-
-export const clearAbi = async (
-  context: HookContext,
-  configEntries: AbiExporterConfig,
-) => {
-  const entries = configEntries.filter((entry) => entry.clear);
-
-  await Promise.all(entries.map((entry) => clearAbiGroup(context, entry)));
-};
 
 export const exportAbi = async (
   context: HookContext,
@@ -27,60 +18,6 @@ export const exportAbi = async (
   await Promise.all(
     configEntries.map((entry) => exportAbiGroup(context, entry)),
   );
-};
-
-const clearAbiGroup = async (
-  context: HookContext,
-  config: AbiExporterConfigEntry,
-) => {
-  const outputDirectory = path.resolve(context.config.paths.root, config.path);
-
-  if (!fs.existsSync(outputDirectory)) {
-    return;
-  }
-
-  // recursively get all files from directory
-
-  const files = (
-    await fs.promises.readdir(outputDirectory, {
-      recursive: true,
-      withFileTypes: true,
-    })
-  )
-    .filter((dirent) => dirent.isFile())
-    .map((dirent) => path.resolve(dirent.parentPath, dirent.name));
-
-  // validate file contents and delete
-
-  await Promise.all(
-    files.map(async (file) => {
-      const contents = await fs.promises.readFile(file, 'utf-8');
-
-      if (path.extname(file) === '.json') {
-        try {
-          // attempt to parse ABI from file contents
-          new Interface(contents);
-        } catch (e) {
-          // file is not an ABI - do not delete
-          return;
-        }
-      } else if (path.extname(file) === '.ts') {
-        if (!contents.includes(TS_TAG)) {
-          // file was not generated by plugin - do not delete
-          return;
-        }
-      } else {
-        // ABIs must be stored as JSON or TS
-        return;
-      }
-
-      await fs.promises.rm(file);
-    }),
-  );
-
-  // delete the directory if it's empty
-
-  await deleteEmpty(outputDirectory);
 };
 
 const exportAbiGroup = async (
